requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
binance-connector==3.4.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dateparser==1.1.8
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.1
orjson==3.9.9
pandas==2.1.1
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
TA-Lib==0.4.28
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
ujson==5.8.0
urllib3==2.0.6
uvicorn==0.23.2
watchfiles==0.21.0
websocket-client==1.6.4
websockets==11.0.3
yarl==1.9.2
